# docs/swagger.yaml
definitions:
  internal_http-server_handlers_actor-movie_delete.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  internal_http-server_handlers_actor-movie_save.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  internal_http-server_handlers_actor_all.Response:
    properties:
      actors:
        items:
          $ref: '#/definitions/postgres.Actor'
        type: array
      error:
        type: string
      status:
        type: string
    type: object
  internal_http-server_handlers_actor_delete.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  internal_http-server_handlers_actor_save.Response:
    properties:
      actor_id:
        type: integer
      error:
        type: string
      status:
        type: string
    type: object
  internal_http-server_handlers_actor_update.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
  postgres.Actor:
    properties:
      actor_id:
        type: integer
      birthdate:
        type: string
      gender:
        type: string
      movies:
        items:
          type: integer
        type: array
      name:
        type: string
    type: object
...
  response.Response:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  search.Response:
    properties:
      actor:
        $ref: '#/definitions/postgres.Actor'
      error:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    email: naklz9@mail.ru
    name: Nikita Zhirnov
    url: https://t.me/belkindelcin
  description: API for VK contest
  termsOfService: http://swagger.io/terms/
  title: Film Library API
  version: "1.0"
paths:
  /actor-movie/delete:
    delete:
      consumes:
      - application/json
      description: Delete actors from movie by movie_id and actors_ids
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: integer
      - collectionFormat: csv
        description: Actors IDs
        in: path
        items:
          type: integer
        name: actors_ids
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http-server_handlers_actor-movie_delete.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete actors from movie
      tags:
      - Actor-Movie
  /actor-movie/save:
    post:
      consumes:
      - application/json
      description: Add actors to movie by movie_id and actors_ids
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: integer
      - collectionFormat: csv
        description: Actors IDs
        in: path
        items:
          type: integer
        name: actors_ids
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http-server_handlers_actor-movie_save.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add actors to movie
      tags:
      - Actor-Movie
  /actor/all:
    get:
      consumes:
      - application/json
      description: Get all actors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http-server_handlers_actor_all.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all actors
      tags:
      - Actor
  /actor/delete:
    delete:
      consumes:
      - application/json
      description: Delete an actor by actor_id
      parameters:
      - description: Actor ID
        in: path
        name: actor_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http-server_handlers_actor_delete.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete an actor
      tags:
      - Actor
  /actor/save:
    post:
      consumes:
      - application/json
      description: Create a new actor by name, gender and birthdate
      parameters:
      - description: Name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Gender
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Birthdate
        in: body
        name: birthdate
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http-server_handlers_actor_save.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new actor
      tags:
      - Actor
  /actor/search:
    get:
      consumes:
      - application/json
      description: Get an actor by actor_id
      parameters:
      - description: Actor ID
        in: path
        name: actor_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/search.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get an actor
      tags:
      - Actor
  /actor/update:
    post:
      consumes:
      - application/json
      description: Update an actor by actor_id
      parameters:
      - description: Actor ID
        in: path
        name: actor_id
        required: true
        type: integer
      - description: Name
        in: body
        name: name
        schema:
          type: string
      - description: Gender
        in: body
        name: gender
        schema:
          type: string
      - description: Birthdate
        in: body
        name: birthdate
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http-server_handlers_actor_update.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update an actor
      tags:
      - Actor
...
swagger: "2.0"